requirements.txt
coco-experiment==2.8.1
cocopp==2.8.0
matplotlib==3.10.7
numpy==2.3.5
scipy>=1.7.0
